+++ client/src/Instance.jsx
@@ -1,15 +1,15 @@
 import { useState, useEffect, useRef } from 'react'
-import Logs from './components/Logs.jsx'
-import ChannelSelect from './components/ChannelSelect.jsx'
-import ModeToggle from './components/ModeToggle.jsx'
-import Controls from './components/Controls.jsx'
-import NewsList from './components/NewsList.jsx'
-import TGSources from './components/TGSources.jsx'
-import FilterSelect from './components/FilterSelect.jsx'
-import AuthorSelect from './components/AuthorSelect.jsx'
-import FiltersTab from './components/FiltersTab.jsx'
-import AuthorsTab from './components/AuthorsTab.jsx'
-import AdminTab from './components/AdminTab.jsx'
+import Logs from './components/Logs'
+import ChannelSelect from './components/ChannelSelect'
+import ModeToggle from './components/ModeToggle'
+import Controls from './components/Controls'
+import NewsList from './components/NewsList'
+import TGSources from './components/TGSources'
+import FilterSelect from './components/FilterSelect'
+import AuthorSelect from './components/AuthorSelect'
+import FiltersTab from './components/FiltersTab'
+import AuthorsTab from './components/AuthorsTab'
+import AdminTab from './components/AdminTab'
 import Button from './components/ui/Button.jsx'
 import apiFetch from './api.js'
