+++ client/src/components/AdminTab/AdminTab.jsx
@@ -1,8 +1,8 @@
 import { useState, useEffect } from 'react'
 import PropTypes from 'prop-types'
-import Button from './ui/Button.jsx'
-import Modal from './ui/Modal.jsx'
-import apiFetch from '../api.js'
+import Button from '../ui/Button.jsx'
+import Modal from '../ui/Modal.jsx'
+import apiFetch from '../../api.js'
 
 const IMAGE_MODELS = ['dall-e-3', 'dall-e-2', 'gpt-image-1', 'gpt-4o']
 // official values from OpenAI docs
