+++ client/src/components/AdminTab/index.js
@@ -1,0 +1,1 @@
+export { default } from './AdminTab.jsx';
