+++ client/src/components/AuthorSelect/index.js
@@ -1,0 +1,1 @@
+export { default } from './AuthorSelect.jsx';
