+++ client/src/components/AuthorsTab/index.js
@@ -1,0 +1,1 @@
+export { default } from './AuthorsTab.jsx';
