+++ client/src/components/ChannelSelect/index.js
@@ -1,0 +1,1 @@
+export { default } from './ChannelSelect.jsx';
