+++ client/src/components/FilterSelect/index.js
@@ -1,0 +1,1 @@
+export { default } from './FilterSelect.jsx';
