+++ client/src/components/FiltersTab/index.js
@@ -1,0 +1,1 @@
+export { default } from './FiltersTab.jsx';
