+++ client/src/components/Logs/index.js
@@ -1,0 +1,1 @@
+export { default } from './Logs.jsx';
