+++ client/src/components/ModeToggle/index.js
@@ -1,0 +1,1 @@
+export { default } from './ModeToggle.jsx';
