+++ client/src/components/NewsItem/NewsItem.jsx
