+++ client/src/components/NewsItem/index.js
@@ -1,0 +1,1 @@
+export { default } from './NewsItem.jsx';
