+++ client/src/components/NewsList/index.js
@@ -1,0 +1,1 @@
+export { default } from './NewsList.jsx';
