+++ client/src/components/SourcesTable/index.js
@@ -1,0 +1,1 @@
+export { default } from './SourcesTable.jsx';
