+++ client/src/components/TGSources/index.js
@@ -1,0 +1,1 @@
+export { default } from './TGSources.jsx';
